cmake: Find qdoc directly instead of relying on qmake path
Qt6ToolsToolsTargets.cmake provides a Qt6::qdoc target, so there is
no need to guess qdoc path based on qmake path. It is the same file
where Qt6::qhelpgenerator is defined, which we use here too.

This fixes build on Debian where we moved Qt6::qmake into a different
path to help with cross-building, leaving Qt6::qdoc where it was.

Change-Id: I9e66ffafc8aeb18edc1fe2b27608caa4ee90c3f7
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>