Remove the use of a QTimer in WindowsStdinReader
When opening a qbs project in QtCreator, it can happen that QtCreator
spawns a lot of qbs processes, depending on the configured kits and build
configurations. The usage of the QTimer to read the stdin on windows then
causes quite the CPU usage for all these processes and processes would
continuously wake up the CPU. This can lead to significant power usage,
which is especially negative on laptops.

This commit prevents this by using Wait routines to either wait for input
or wait for a signaled exit event. This exit event is signaled in the
destructor of the reader, where it also waits until the thread is
actually finished.

Because stdin can be a number of different "file" types with each acting
differently regarding wait routines and reads, specific readers are
necessary for the different "file" types.

Change-Id: I14d04414c98c6d0dfb93185cc7aaf2cd64c2b0b6
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>