Loader: Do not error out when encountering duplicate multiplex values
In qbs, values of list properties specified in a project file are merged
with those from the profile.
As of fb52fed84a1510a7de0172e643d6fd66a780e2e8, this is also true for
the special multiplexing properties ("qbs.architectures",
"qbs.buildVariants" etc). This means it can now happen that we end up
with duplicate entries in these lists implicitly. Therefore, it no longer
makes sense to throw an error when such duplicates are encountered, as
they were not necessarily specified verbatim. Instead, we simply ignore
the duplicate value and carry on normally.

Change-Id: Ie4fed2081bebd2b0dd62aa873cafed769b308e97
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>