android: remove old Qt versions from CI
Ideally, we need to rebuild all images when docker file is changed and
maintaning that amount of images is quite a burden.

Also, most users use LTS versions like 5.15, 6.2, 6.5 or the latest Qt
version.

Leave 6.3 and 6.4 for now in order to test regressions since there were
some breaking changes in these versions.

Change-Id: Id4f7e817119f3226c1a2350d0e6f4772da162943
Reviewed-by: Raphaël Cotty <raphael.cotty@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>