Loader: Sort products before resolving
We'd like to maximize the number of products that finish resolving at
the first attempt, so we sort them initially by the number of Depends
items, which acts as a proxy for the number of actual dependencies,
which is unknown a priori. At the very least, this guarantees that
products with no dependencies get handled first.
Numbers:
  - When resolving qbs itself, the number of deferrals goes down
    from 103 to 61, with a total of 54 products.
  - When resolving Qt Creator, the number of deferrals goes down
    from 1065 to 760, with a total of 304 products.
Our benchmarking tool shows no effect, but this is still the right thing
to do out of principle. In particular, with parallelized resolving it
will help avoid starting threads unnecessarily.

Change-Id: I0af363833df5711b8148bbd98491b9379cb68a42
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>