Loader: Ensure ModuleLoader only returns Module items
Until now, it could also return (pseudo-)module instances, which was
unexpected at the calling side. Fix that by moving the code that deals
with a module not being found to DependenciesResolver entirely. This
cleanly separates the responsibilities and also improves an error
message, as a formerly inadvertently dead code path is now being taken.

Change-Id: I2f2e3051550981885a7073fde271b9686db00c6a
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>