apple: fix some DMG module issues
Fix keys used in settings to match those that are actually
used in python.

Removed Korean and Russian license agreements - they are not
listed in the allowed languages.

Also, even when added to the list, Russian eula.txt makes hdutil to
return with an error for some reason.

Those need further investigation.

Fixes: QBS-1753
Change-Id: I510d4f1896625d6a0b63c4005288d29069776c9a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>