Remove dependency of SourceWildCards on ResolvedGroup
SourceWildCards had an unnecessary dependency on ResolvedGroup which
could be a historical relic. However, it is functionally independent and
can be useful in other contexts as well, for example wildcard project
references. We really do not want to duplicate the complicated wildcard
expansion code.

This patch is a first step in a refactoring series.

Task-number: QBS-1749
Change-Id: I2ce0e784d4668f7cee686cd4c2101aef3ce38eda
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>