Add Rust-like Mutex
This change implements a Rust-like Mutex class
that prevents unlocked access to the underlying data.

As seen recently, it is easy to make a mistake and lock
the wrong mutex. Also, this prevents micro-optimizations
where shared state is accessed without locking corresponding
mutex in hope that there is some other barrier.

Change-Id: I52c7a2fa68cc97be07a490c35aa76fe717032fdf
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>