qbspkgconfig: Add support for the definePrefix option
On windows/msys, it is crucial to use autodetection for package
paths - otherwise, Qbs is not able to detect e.g. protobuf
package.

The pkg-config has the --define-prefix switch that overrides
the ${prefix} variables on the package with one based on
the .pc file locaiton. Implement similar option and set it
to true on Windows.
Also fix setting sysroot to avoid duplication of sysroot
in paths when prefix is autodetected.

Change-Id: I2754e48a96cb2e5f01ecace616188782cb5b0d7a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>