DependenciesResolver: Don't attach properties on non-present modules
We used to do that anyway and then suppress the resulting exception, but
that can lead to an unexpected state later. The sequence that triggered
the linked bug was as follows:
  - load module via non-required dependency in other module
  - exchange placeholder item in the product item and try to attach
    non-existing property on the module instance
  - suppress the resulting exception (because the dependency
    was not required) and continue
  - load module again (required this time) on product level
  - exchange placeholder item again -> assertion, because we have a
    module instance instead of the expected placeholder
Another way to fix this would have been to revert exchanging the
placeholder item after encountering an error, but that would be more
complex, and there is no reason to attach properties to non-existing
modules anyway.

Fixes: QBS-1776
Change-Id: I581a076d1d872616b186e4015873baeea211b647
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>