Evaluator: Catch infinite recursion in property evaluation
This patch was originally committed as
386be8cc4020d7f87d6879cd8c22c6842b49d268.
We reverted it in c99763b462a93b547f4f3f41ab84e9e793276d4c, because it
broke the Qt Ceator build. We then identified the actual culprit for that
and fixed it in 410653d331469aacef5db7568132a04cd92a9302, so this one can
now be re-applied.

Fixes: QBS-1793
Change-Id: I5efec09d08291892c3840e479f48b342f633d30e
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>