GitHub actions: fix windows job
The problem was that VM images updated the compiler
version to 14.40 which is not supported prior to
Conan 2.4.1.

Use the github action to install Conan as Choco still
has 2.4.0.

Also, separate conan profiles for proto/capnroto and
provider tests - we still need to specify the older
compiler to be able to download prebuilt packages.

Change-Id: I8f8973ddb1ab7d1db9c60d7e3f7deb8643740843
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>