Fix "implicit else" behavior of Properties item
The fact that a Properties item caused the referenced module properties
to be altered even when its condition is false was both unintuitive and
inconvenient.
We now continue looking in the next-highest-prioritized value when
encountering a JS value with no enabled alternatives and no base value,
which results in the expected behavior.
Also remove most usages of "original" from our own modules, as they are
no longer needed.

Fixes: QBS-1254
Change-Id: Ie46189710de5d79a087c9e7a0341f9c32340f61c
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>