Add support for include Android resouces from dependencies
When building against Java library packages that include resources,
those resources need to be combined with the resources from the
application.

This change adds an Android.sdk.extraResourcePackages property that can
be used to specify the list of packages to include when compiling the
Android resources. The extra package names are passed to the
'aapt2 link' using the --extra-packages option.

Change-Id: I6fe5e459787ba497a772289d8fce84902bc6e227
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>