Evaluator: Fix resolving path properties
Do not use a Project item's sourceDirectory property as the base dir, as
that one always points to the project root.
(More generally, it appears that path properties of Project items are
not well-defined, but that's out of this patch's scope.)
Amends 9662cf41aeb8ccc0f3ad76d7236460cf31356c03.

Change-Id: Iacdab29d8e9e926a301292755e5c1b822c0df10e
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>