ModuleInstantiator: Update ids in scope
Early in the process of setting up product and module instance items,
the ids of all child items are inserted into the product/module scope so
that they can be accessed from anywhere in the item.
However, ids in Depends items are special, as they are supposed to refer
to the loaded module rather than the Depends item itself.
Therefore, when instantiating a module, we need to update the entry so
that it now points to the module instance.

Change-Id: I3b028a2aa19a4f1492392c49147d6c6f88d5462a
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>