Loader: Check version already when loading a module
Otherwise, if two variants of a module are available in different search
paths, with only the second one fulfilling the version requirement, the
first one will get loaded and then later fail the version check, causing
project resolving to fail.
This currently happens in practice with projects requiring Qt 6 and is
quite annoying.

Change-Id: I7b2f14162c039b487040a8188f1ba5eeda7459c0
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>