Qt support: Remove change tracking support for exotic moc feature
For a header file myclass.h containing a class with a Q_OBJECT macro, moc
creates a cpp file moc_myclass.cpp. This file is usually compiled into an
object file which then gets linked into a library or application.
However, there is the little-known feature that one can also #include
moc_myclass.cpp in a source file, in which case it must not get compiled.
To support this, our moc rule scans all cpp files in the product and
assigns either the "hpp" or the "cpp" tag to moc_myclass.cpp, depending
on whether it gets included or not.
As a side effect of using auxiliaryInputs wrong, qbs used to be able to
detect when such an include existed but then got removed and vice versa.
When we fixed the use of auxiliaryInputs in
0c3e8667105302ba05f4352808e4185733105f05, we kept that feature alive by
making the moc rule use explicitlyDependsOn instead. However, that
inadvertantly led to moc getting re-run on all headers if any cpp file
was touched, which is obviously not acceptable.
We could keep the automatic change tracking by introducing a fourth
input-like rule property that has the semantics of re-running the rule
scripts (as opposed to the transformers) when files with matching tags
are touched. However, I do not think that is worth doing just to support
a workflow that almost never occurs in practice. Users can easily touch
the header file after adding/removing such an #include directive.
Adapting the test case uncovered a spurious assertion that was also
fixed.

Fixes: QBS-1823
Change-Id: I857a2c4d436ad4b5caacd67ab72ca75d41a10704
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>