Evaluator: Improve error handling
As we cannot throw C++ exceptions across QuickJS, we kept retrieving and
re-throwing JS exceptions, which lost structural information and also
messed with the error locations.
Instead, we now keep the original exception stored in our script engine
and throw it as soon as we are out of JavaScript land.

Change-Id: I3bc434191726410eea91fad5ff32073627e29019
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>