Add archiverFlags property to cpp module
All toolchain offer tools to generate static libraries. This is usually
not done by a linker, but by the so-called archiver. We should not mix
up the terms linking and archiving. And since we already have flags for
all tools and we have properties like archiverName, archiverPath, it
feels only natural to also add an archiverFlags property.

Task-number: QBS-1832
Change-Id: Ide0609616be50a20eb59c9af04afc61b36fc45f2
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>