Executor: Fix artifact rescuing
Before potentially re-running a transformer's commands, we first rescue
its artifact data, if applicable. This rescue data contains information
retrieved during execution of the transformer's commands and prepare
script. At rescue time, the prepare script might have re-run, in which
case the current information in the transformer is up to date and must
not be overwritten with outdated rescue data. Consider that in
Transformer::rescueFromArtifactData().

Fixes: QBS-1854
Change-Id: Ib32f2428768a57ab0df3453313b16076757d7c91
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>