gcc: handle absolute module source paths in modules.json files
At least on Debian (and derivative), the JSON file with std modules has
absolute paths:

  $ gcc -print-file-name=libstdc++.modules.json
  /usr/lib/gcc/x86_64-linux-gnu/15/libstdc++.modules.json
  $ cat /usr/lib/gcc/x86_64-linux-gnu/15/libstdc++.modules.json
  {
    "version": 1,
    "revision": 1,
    "modules": [
      {
        "logical-name": "std",
        "source-path": "/usr/include/c++/15/bits/std.cc",
        "is-std-library": true
      },
      {
        "logical-name": "std.compat",
        "source-path": "/usr/include/c++/15/bits/std.compat.cc",
        "is-std-library": true
      }
    ]
  }

Joining the directory of the JSON file with the source path results in
an invalid path, thus not finding & using the std modules.

To handle this situation, first check whether the source path is an
absolute path, returning it as-is.

Change-Id: Ia5177235d8addc3ad621d520c6e4522d11ac9c0b
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>