probe: Allow to use ToolchainInstallInfo in an unordered set containers
Previously we are used the std::vector<ToolchainInstallInfo> for storing
the detected toolchains (e.g. from the system paths or from the registry).

This forced us to make an additional actions to filter out a duplicates
before sorting and so on, that complicates an implementation in case we
have more than two vectors.

Right now we can use the std::unordered_set<ToolchainInstallInfo> which
saves us in the future from this extra work, and simplifies a code.

Change-Id: If8c4baab89a87f0e56e64ac0df96f457cdda43f2
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>