tests: Add missing 'consoleApplication' property to blackbox tests
... for the test applications that have the main() entry.

This is a more correct way, because it matters for some compilers on
Windows (for example, for Borland C++, Open Watcom, etc.). Because
they requires a different startup code for the console and the win32
applications, where the win32 applications uses the WinMain() entry
instead of the main() entry.

Thus, this fixes the linking of test applications for these toolchains,
and allows to use also these tests for testing an other toolchains.

Change-Id: I126949fc0b55fe44f8995196866c141691843969
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>